transmit silence first
this is how libhackrf does it. we don't fill the initial
buffers, therefore transferring
transferCount * transferBufferSize / 2 samples of silence.

the advantage is, for a heavy callback, we make sure the
buffer is filled just after enabling the transceiver, which
prevents *most* drops at the start of an OUT transfer

there's a better solution (fill those buffers before even
starting the transceiver), but I'll implement it later